+++ lang/en/theme_campus.php
@@ -142,6 +142,12 @@ $string['textcolourdesc'] = 'Set the text colour.';
 $string['linkcolour'] = 'Link text colour';
 $string['linkcolourdesc'] = 'Set the link text colour.';
 
+$string['iconcoloursetting'] = 'Use icon colour setting';
+$string['iconcoloursetting_desc'] = 'Use the icon colour setting for the icons.  The icon colour functionality is new.  If you experience problems with it, then turn it off and run \'grunt svg\' on the Node.js command prompt.  Please see \'Gruntfile.js\' for full details.';
+
+$string['iconcolour'] = 'Icon colour';
+$string['iconcolour_desc'] = 'The colour for the icons.';
+
 $string['headingcolour'] = 'Heading colour';
 $string['headingcolourdesc'] = 'Set the heading colour.';