+++ Changes.md
@@ -3,6 +3,7 @@ Version Information
 25th July 2016 - Version 3.1.1.2
   1. Setting coursepagepageheadinglocation/frontpagepageheadinglocation is not always applied correctly - task #863.
   2. Userpicture in navbar: style improvements - task #861.
+  3. Color of block icons - task #859.
 
 23rd July 2016 - Version 3.1.1.1
   1. Hamburger button looks ugly in many cases - task #860.