+++ Changes.md
@@ -1,5 +1,8 @@
 Version Information
 ===================
+28th July 2016 - Version 3.1.1.3
+  1. Forumpost font colour - task #862.
+
 25th July 2016 - Version 3.1.1.2
   1. Setting coursepagepageheadinglocation/frontpagepageheadinglocation is not always applied correctly - task #863.
   2. Userpicture in navbar: style improvements - task #861.